requirements.txt
requests
Pillow
pyautogui
tk
comtypes
python-dotenv
